Rework non_account_contract e2e test
